fixes https://app.asana.com/0/722421058580118/728842614993042
The [app_schedule worker="5"] was not working correctly a it was always getting the $args['workers'][0] resulting on a wrong schedule list.

If there's a worker_id set ( i.e. 5 ) it will now read his schedule instead. And if not it will just revert back to [0].

Also some CS just on the spots I touched.